==> 03_Skripte/1_Datenbereinigung.ipynb <==
# 1. Datenbereinigung

## 1. Bibliotheken importieren

import pandas as pd   # pandas für Datenanalyse
import numpy as np    # NumPy für numerische Operationen
import os             # os für Dateipfade und Ordneroperationen

## 2. Importierung den Datensatz

df_prods = pd.read_csv(r'C:\Users\faink\Desktop\Achievement_4\02_Daten\Rohe_Daten\products.csv', index_col = False)

## 3. Erste 20 Zeilen

df_prods.head(20)

## 4. Letzte 35 Zeilen

df_prods.tail(35)

## 5. Die Namen der Spalten

df_prods.columns

## 6. Die Anzahl der Zeilen und Spalten

df_prods.shape

## 7. Der maximale Wert der Spalte „aisle_id“ 

df_prods.describe()

die maximale Wert der Spalte „aisle_id“ ist gleich 134.000000

## 8. Der Datentyp der Spalte „department_id“

df_prods["department_id"].dtype

Der Datentyp der Spalte „department_id“ ist gleich int64
