==> 03_Skripte/2_Subsetting & neue DataFrames.ipynb <==
# 2. Subsetting & neue DataFrames

## 1. Bibliotheken importieren

import pandas as pd   # pandas für Datenanalyse
import numpy as np    # NumPy für numerische Operationen
import os             # os für Dateipfade und Ordneroperationen

## 2. Importierung den Datensatz

df_prods = pd.read_csv(r'C:\Users\faink\Desktop\Achievement_4\02_Daten\Rohe_Daten\products.csv', index_col = False)

## 3. Format einer Identifikationsvariable ändern

df_ords = pd.read_csv(r'C:\Users\faink\Desktop\Achievement_4\02_Daten\Rohe_Daten\orders.csv', index_col = False)

# Umwandlung von order_id in String
df_ords['order_id'] = df_ords['order_id'].astype(str)

# Alternativ: user_id umwandeln
df_ords['user_id'] = df_ords['user_id'].astype(str)

# Überprüfung
df_ords.dtypes

Die Variable wurde in den Datentyp „String“ umgewandelt, da es sich um eine Identifikationsvariable handelt und keine numerischen Berechnungen durchgeführt werden sollen.

## 4. Eine nicht intuitive Variable umbenennen (ohne Überschreiben des DataFrames)

df_ords_renamed = df_ords.rename(columns={'order_dow': 'order_day_of_week'})

# Überprüfung
df_ords_renamed.columns

Die Variable wurde umbenannt, um die Lesbarkeit und Verständlichkeit des Datensatzes zu verbessern.

## 5. Die meistgenutzte Bestellstunde ermitteln

df_ords['order_hour_of_day'].value_counts().sort_index()

Die am stärksten frequentierte Stunde:

df_ords['order_hour_of_day'].value_counts().idxmax()


Die Analyse zeigt, zu welcher Tageszeit die meisten Bestellungen aufgegeben wurden.

## 6. Subset für Frühstücksprodukte erstellen

df_breakfast = df_prods.loc[df_prods['department_id'] == 14]

df_breakfast.head()


Es wurde ein Subset erstellt, das ausschließlich Produkte aus der Frühstückskategorie enthält.

## 7. Subset für „Dinner Party“-Produkte erstellen

df_dinner_party = df_prods.loc[
    df_prods['department_id'].isin([5, 13, 7, 15])
]

df_dinner_party.head()

Dieses Subset enthält Produkte, die typischerweise für eine Dinner-Party relevant sind (z. B. Alkohol, Feinkost, Getränke sowie Fleisch/Meeresfrüchte).

## 8. Anzahl der Zeilen im neuen DataFrame prüfen

df_dinner_party.shape

Die Anzahl der Beobachtungen im neu erstellten Subset wurde überprüft.

## 9. Alle Daten für user_id = 1 anzeigen

df_user_1 = df_ords[df_ords['user_id'] == '1']

df_user_1.head()


Es wurde ein Subset erstellt, das alle Bestellungen des Nutzers mit der ID 1 enthält.

## 10. Deskriptive Statistik für diesen Nutzer

df_user_1.describe()

Die deskriptive Analyse liefert Informationen über das Bestellverhalten des ausgewählten Nutzers (z. B. bevorzugte Uhrzeit oder Bestellintervalle).

## 11. Export von df_ords

path = r'C:\Users\faink\Desktop\Achievement_4\02_Daten'

df_ords.to_csv(os.path.join(path, 'Prepared_Daten', 'orders_wrangled.csv'), index=False)



## 12. Struktur des Notebooks (Markdown-Überschriften)

# Data Wrangling

## Formatierung von Variablen

## Umbenennung von Variablen

## Subsetting von Daten

## Analyse eines einzelnen Nutzers

## Export der bereinigten Datensätze
