==> 03_Skripte/6_Gruppierung & Aggregation.ipynb <==
# 6. Gruppierung & Aggregation

## 1. Bibliotheken importieren und Daten laden

# Bibliotheken importieren
import pandas as pd
import os

# Pfade definieren
base_path = r"C:\Users\faink\Desktop\Achievement_4\02_Daten"
raw_path = os.path.join(base_path, "Rohe_Daten")
prepared_path = os.path.join(base_path, "Prepared_Daten")

# DataFrame laden (angenommen als Pickle-Datei aus der vorherigen Übung)
ords_prods_merge = pd.read_pickle(os.path.join(prepared_path, "ords_prods_merge.pkl"))

# products.csv laden (enthält product_id, product_name, aisle_id, department_id, evtl. price)
products = pd.read_csv(os.path.join(prepared_path, "products_checked.csv"))

# Überblick über die Daten
ords_prods_merge.head()

## 2. Aggregierter Mittelwert von „ordernumber“ pro „departmentid“ (gesamter DataFrame)

# Department-Datei laden
departments = pd.read_csv(os.path.join(prepared_path, "departments_wrangled.csv"))

# Überblick über die Daten
departments.head()

## 3. Loyalitäts-Flag erstellen

# Loyalitäts-Flag erstellen
ords_prods_merge['loyalty_flag'] = ords_prods_merge.groupby('user_id')['order_number'] \
    .transform(lambda x: 'Loyal Customer' if x.max() > 40 else ('Regular Customer' if x.max() > 10 else 'New Customer'))

# Prüfen
ords_prods_merge[['user_id', 'order_number', 'loyalty_flag']].head(10)

## 4. Produktpreis-Statistiken nach Loyalität

ords_prods_merge.head()

## 5. Ausgaben-Flag erstellen

products.head()

# Mit ords_prods_merge zusammenführen
ords_prods_merge = ords_prods_merge.merge(products[['product_id','prices']], on='product_id', how='left')

# Prüfen
ords_prods_merge[['user_id','product_id','prices']].head()

# Durchschnittlicher Produktpreis pro Nutzer
avg_price_per_user = ords_prods_merge.groupby('user_id')['prices'].mean().reset_index()
avg_price_per_user['spending_flag'] = avg_price_per_user['prices'].apply(lambda x: 'Low spender' if x < 10 else 'High spender')

# DataFrame zusammenführen
ords_prods_merge = ords_prods_merge.merge(avg_price_per_user[['user_id','spending_flag']], on='user_id', how='left')

# Prüfen
ords_prods_merge[['user_id','prices','spending_flag']].head(10)


## 6. Bestellhäufigkeits-Flag erstellen

# Median days_since_prior_order pro Nutzer
median_days = ords_prods_merge.groupby('user_id')['days_since_prior_order'].median().reset_index()

def frequency_flag(days):
    if days > 20:
        return 'Nicht-häufiger Kunde'
    elif days > 10:
        return 'Regelmäßiger Kunde'
    else:
        return 'Häufiger Kunde'

median_days['frequency_flag'] = median_days['days_since_prior_order'].apply(frequency_flag)

# DataFrame zusammenführen
ords_prods_merge = ords_prods_merge.merge(median_days[['user_id','frequency_flag']], on='user_id', how='left')

# Prüfen
ords_prods_merge[['user_id','days_since_prior_order','frequency_flag']].head(10)

## 7. DataFrame exportieren 

# Als Pickle-Datei speichern
ords_prods_merge.to_pickle(os.path.join(prepared_path, "ords_prods_merge_prepared.pkl"))
